chore: postpone refactoring until login feature is complete
