+++ lib/domain/board/recruit/entities/recruit_detail_entity.dart
