+++ lib/domain/board/recruit/repositories/recruit_detail_repository.dart
@@ -1,4 +1,4 @@
-import 'package:dongsoop/domain/board/recruit/entities/detail/recruit_detail_entity.dart';
+import 'package:dongsoop/domain/board/recruit/entities/recruit_detail_entity.dart';
 import 'package:dongsoop/presentation/board/common/enum/recruit_types.dart';
 
 abstract class RecruitDetailRepository {
