+++ lib/domain/board/recruit/use_cases/recruit_detail_use_case.dart
@@ -1,5 +1,5 @@
-import 'package:dongsoop/domain/board/recruit/entities/detail/recruit_detail_entity.dart';
-import 'package:dongsoop/domain/board/recruit/repositories/detail/recruit_detail_repository.dart';
+import 'package:dongsoop/domain/board/recruit/entities/recruit_detail_entity.dart';
+import 'package:dongsoop/domain/board/recruit/repositories/recruit_detail_repository.dart';
 import 'package:dongsoop/presentation/board/common/enum/recruit_types.dart';
 
 class RecruitDetailUseCase {
