+++ lib/presentation/board/recruit/detail/providers/recruit_detail_use_case_provider.dart
@@ -1,6 +1,6 @@
 import 'package:dongsoop/core/providers/dio_provider.dart';
-import 'package:dongsoop/data/board/recruit/repositories/detail/recruit_detail_repository_impl.dart';
-import 'package:dongsoop/domain/board/recruit/use_cases/detail/recruit_detail_use_case.dart';
+import 'package:dongsoop/data/board/recruit/repositories/recruit_detail_repository_impl.dart';
+import 'package:dongsoop/domain/board/recruit/use_cases/recruit_detail_use_case.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 final recruitDetailUseCaseProvider = Provider<RecruitDetailUseCase>((ref) {
