+++ lib/presentation/board/recruit/detail/view_models/recruit_detail_view_model.dart
@@ -1,5 +1,5 @@
-import 'package:dongsoop/domain/board/recruit/entities/detail/recruit_detail_entity.dart';
-import 'package:dongsoop/domain/board/recruit/use_cases/detail/recruit_detail_use_case.dart';
+import 'package:dongsoop/domain/board/recruit/entities/recruit_detail_entity.dart';
+import 'package:dongsoop/domain/board/recruit/use_cases/recruit_detail_use_case.dart';
 import 'package:dongsoop/presentation/board/common/enum/recruit_types.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
